pyproject.toml
[project]
name = "pydantic-ai-tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "prettytable>=3.14.0",
    "pydantic-ai>=0.0.24",
    "python-dotenv>=1.0.1",
]
